UGA Hacks Game API/requirements.txt
fastapi 
uvicorn
google-genai
python-dotenv
